Info.md
﻿# Cosmos
> The universe of algorithm and data structures, includes most of the data structures that are required by coders.

**Cosmos** is your personal offline collection of every algorithm and data structure one will ever encounter and use in a lifetime. This provides solutions in a various languages spanning `C`, `C++`, `Java`, `Python`, `Go` and others. 

This work is maintained by a community of hundreds of people and is a _massive collaborative effort_ to bring the readily available coding knowledge **offline**. 

# Cosmic Structure

Following is the high-level structure of cosmos:
* Backtracking	
* Bit manipulation
* Computational geometry
* Data structures
* Divide conquer
* Dynamic programming
* Graph algorithms
* Greedy algorithms
* Mathematical algorithms	
* Randomized algorithms	
* Search 
* Sorting	
* String algorithms
* Unclassified 👻 

Each type has several hundreds of problems with solutions in several languages spanning `C`, `C++`, `Java`, `Python`, `Go` and others.

# Maintainers

This is a massive collaboration and to keep the quality intact and drive the vision in the proper direction, we have maintainers.

> Maintainers are your friend forever. They are vastly different from moderators. 

The task of maintainers is to review pull requests, suggest further quality additions and keep the work up to date with the current state of the world. 🌍 

[Check out our current maintainers](https://github.com/OpenGenus/cosmos/blob/master/.github/CONTRIBUTING.md)

Let us know that you would like to be a maintainer in the Slack channel *#algorithms* and we will review and add you upon subsequent contributions. [Join us at Slack](http://slack.opengenus.org/).

# Contributions

The success of our vision to bring knowledge offline** depends on you. Even a small contribution helps. All forms of contribution are highly welcomed and valued. 

When you contribute, your name with a link (if available) is added to our [contributors list](https://github.com/OpenGenus/cosmos/wiki/contributors).

# License

We believe in freedom and improvement. [GNU General Public License v3.0](https://github.com/AdiChat/cosmos/blob/master/LICENSE)